Delete low-value helper components
I'd rather reserve the package for components that have proven repeatedly useful, like `Classes` and `HTML5`.
